style: 성별, 목표라벨 상수폴더로 이동
