+++ src/types/profile.ts
@@ -3,20 +3,10 @@ export enum Goal {
   BalancedDiet = "balanced_diet",
 }
 
-export const GoalLabels: Record<Goal, string> = {
-  [Goal.WeightLoss]: "체중감량",
-  [Goal.BalancedDiet]: "균형식단",
-};
-
 export enum Gender {
   Male = "male",
   Female = "female",
 }
-
-export const GenderLabels: Record<Gender, string> = {
-  [Gender.Male]: "남성",
-  [Gender.Female]: "여성",
-};
 
 export interface ProfileData {
   id?: string;
